Hice el servidor con express y lo concecte con el front, todo documentado en server.js. No cambien las rutas ni direccion de los archivos
