--- css/carrito.css
@@ -1,234 +1,0 @@
-body.oscuro {
-  background-color: #121212;
-  color: #f5f5f5;
-}
-
-body.oscuro header {
-  background-color: #1e1e1e;
-}
-
-body.oscuro .btn-vaciar,
-body.oscuro button {
-  background-color: #333;
-  color: #f5f5f5;
-  border: 1px solid #555;
-}
-
-body.oscuro .modal-contenido {
-  background-color: #222;
-  color: #fff;
-}* 
-
-{
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: 'Poppins', sans-serif;
-}
-
-body {
-  background: linear-gradient(135deg, #0f0f0f, #1b1b1b);
-  color: #f0f0f0;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-}
-
-header {
-  background: #1e1e1e;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 15px 40px;
-  border-bottom: 1px solid #2a2a2a;
-}
-
-.logo {
-  width: 200px;
-  border-radius: 10px;
-}
-
-nav button {
-  background: linear-gradient(90deg, #1f6feb, #0056d6);
-  border: none;
-  color: white;
-  padding: 10px 18px;
-  border-radius: 8px;
-  cursor: pointer;
-  margin-left: 10px;
-  transition: 0.3s;
-}
-
-nav button:hover {
-  transform: translateY(-2px);
-}
-
-main {
-  flex: 1;
-  padding: 40px;
-  text-align: center;
-}
-
-.carrito-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 15px;
-  margin-top: 30px;
-}
-
-.item-carrito {
-  background: #1a1a1a;
-  border-radius: 10px;
-  padding: 15px;
-  width: 90%;
-  max-width: 600px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.item-carrito img {
-  width: 60px;
-  border-radius: 8px;
-}
-
-.item-carrito h3 {
-  flex: 1;
-  margin: 0 15px;
-  text-align: left;
-}
-
-.controles {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-
-.btn-eliminar {
-  background: #ff4747;
-  border: none;
-  color: white;
-  padding: 8px 12px;
-  border-radius: 6px;
-}
-
-.btn-eliminar:hover {
-  background: #d83838;
-}
-
-.btn-vaciar {
-  margin-top: 20px;
-  background: #ff4747;
-  border: none;
-  color: white;
-  padding: 10px 18px;
-  border-radius: 8px;
-}
-
-.total-section {
-  margin-top: 30px;
-  text-align: center;
-  font-size: 1.3rem;
-}
-
-/* Estilo para el logo */
-h1 {
-    margin-left: -200px;         
-    display: block;
-        
-}
-
-/* Si querés alinearlo al centro en la parte superior */
-button {
-  background-color: transparent;
-  border: none; 
-  color: inherit; 
-  cursor: pointer; 
-}
-
-/* ====== MODAL DE PAGO ====== */
-.modal {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(15, 15, 15, 0.9);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-}
-
-.modal.oculto {
-  display: none;
-}
-
-.modal-contenido {
-  background: #1a1a1a;
-  padding: 30px 40px;
-  border-radius: 15px;
-  box-shadow: 0 0 20px #0056d6;
-  text-align: center;
-  width: 90%;
-  max-width: 400px;
-  color: #f0f0f0;
-  animation: aparecer 0.3s ease-out;
-}
-
-.modal-contenido h2 {
-  margin-bottom: 15px;
-  color: #1f6feb;
-}
-
-.modal-contenido input {
-  width: 80%;
-  padding: 10px;
-  border-radius: 8px;
-  border: 1px solid #2a2a2a;
-  background: #0f0f0f;
-  color: white;
-  font-size: 1rem;
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-.modal-botones {
-  display: flex;
-  justify-content: space-around;
-}
-
-.modal-botones button {
-  background: linear-gradient(90deg, #1f6feb, #0056d6);
-  border: none;
-  color: white;
-  padding: 10px 20px;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: 0.3s;
-}
-
-.modal-botones button:hover {
-  transform: translateY(-2px);
-}
-
-#btnCancelar {
-  background: #ff4747;
-}
-
-#btnCancelar:hover {
-  background: #d83838;
-}
-
-@keyframes aparecer {
-  from {
-    opacity: 0;
-    transform: scale(0.9);
-  }
-  to {
-    opacity: 1;
-    transform: scale(1);
-  }
-}
