+++ package.json
@@ -1,0 +1,27 @@
+{
+  "name": "trabajo-practico-progra3_lautaromoro-ignaciofernandez",
+  "version": "1.0.0",
+  "description": "1: Creamos dos archivos CSS: uno para index.html y otro para productos.html. También se buscó una API para el catálogo de ropa y se crearon los respectivos archivos JS: api.js y productos.js\r 2: Se eligio el tipo de ropa para la pagina y se cambiaron ciertos estilos de los .html\r 3: Se crearon dos archivos HTML: carrito.html y ticket.html, a su vez se le hicieron sus respectivos archivos JS para estos ultimos: carrito.js y ticket.js\r 4: Se le dió nombre y se creó un logo a la página\r 5: Se cambió el estilo y se crearon un mini menu en productos.html con las diferentes categorias de ropa a elegir(sin funcionalidad hasta el momento).",
+  "main": "server.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node server.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/LautaroMoro/Trabajo-Practico-Progra3_LautaroMoro-IgnacioFernandez.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/LautaroMoro/Trabajo-Practico-Progra3_LautaroMoro-IgnacioFernandez/issues"
+  },
+  "homepage": "https://github.com/LautaroMoro/Trabajo-Practico-Progra3_LautaroMoro-IgnacioFernandez#readme",
+  "type": "module",
+  "dependencies": {
+    "axios": "^1.13.2",
+    "cors": "^2.8.5",
+    "express": "^5.1.0"
+  }
+}
