--- productos.html
@@ -1,56 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Catálogo de Productos - TP1</title>
-  <link rel="stylesheet" href="css/productos.css">
-  <script src="js/api.js" defer></script>
-  <script src="js/productos.js" defer></script>
-</head>
-
-<body>
-  <header>
-    <h1>
-      <button id="logoBtn"><img src="nombre.png" alt="Nombre Aplicación"></button>
-    </h1>
-    <div class="menu">
-      <button class="menu-button">Categorías</button>
-      <select id="categoriaSelect" class="dropdown">
-        <option value="all">Mostrar todos</option>
-        <option value="mens-shirts">Remeras</option>
-        <option value="womens-dresses">Vestidos</option>
-        <option value="mens-shoes">Zapatillas</option>
-      </select>
-    </div>
-
-    <nav>
-      <a href="carrito.html">Carrito 🛒</a>
-      <button id="adminLoginBtn">Login Admin</button>
-    </nav>
-  </header>
-
-  <main>
-    <section id="productos-container" class="productos-grid">
-      <p>Cargando productos...</p>
-    </section>
-
-    <!-- Paginación -->
-    <div id="pagination" style="text-align:center; margin: 20px;">
-      <button id="prevPage">Anterior</button>
-      <span id="pageInfo"></span>
-      <button id="nextPage">Siguiente</button>
-    </div>
-  </main>
-
-  <footer>
-    <p>Trabajo Practico. Conoce a nuestro equipo!!
-      <a href="https://github.com/LautaroMoro/Trabajo-Practico-Progra3_LautaroMoro-IgnacioFernandez.git" target="_blank">
-        <img src="logoGITHUB.png" alt="GitHub" width="25" height="25">
-      </a>
-    </p>
-  </footer>
-
-  </div>
-</body>
-</html>
