+++ public/css/bienvenida.css
@@ -1,0 +1,115 @@
+/* Reset y tipografía */
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: 'Poppins', sans-serif;
+}
+
+body {
+    background: linear-gradient(135deg, #1f1f1f, #121212);
+    color: #f5f5f5;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-height: 100vh;
+    padding: 20px;
+}
+
+/* Animación bienvenida */
+.welcome-animation {
+    position: absolute;
+    font-size: 3rem;
+    font-weight: bold;
+    text-align: center;
+    opacity: 1;
+    transition: all 0.8s ease;
+}
+
+/* Contenido principal que aparece después */
+.main-content {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    opacity: 0;
+}
+
+.main-content.fade-in {
+    animation: fadeIn 1s forwards;
+}
+
+@keyframes fadeIn {
+    from {opacity: 0; transform: translateY(20px);}
+    to {opacity: 1; transform: translateY(0);}
+}
+
+/* Logo y títulos */
+.logo {
+    width: 140px;
+    margin-bottom: 20px;
+}
+
+h1 {
+    font-size: 2rem;
+    margin-bottom: 30px;
+}
+
+/* Formulario nombre */
+.name-section {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 12px;
+    margin-bottom: 30px;
+}
+
+input[type="text"] {
+    padding: 12px 15px;
+    width: 250px;
+    border-radius: 10px;
+    border: none;
+    background-color: #2a2a2a;
+    color: #fff;
+}
+
+button {
+    padding: 10px 20px;
+    background: linear-gradient(90deg, #1f6feb, #0056d6);
+    border: none;
+    border-radius: 10px;
+    color: #fff;
+    cursor: pointer;
+    transition: 0.3s;
+}
+
+button:hover {
+    background: linear-gradient(90deg, #155ecf, #0041a3);
+    transform: translateY(-2px);
+}
+
+/* Quiénes somos */
+.about-section {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 10px;
+}
+
+#aboutText {
+    max-width: 400px;
+    text-align: center;
+    color: #ccc;
+    opacity: 0;
+}
+
+#aboutText.fade-in {
+    animation: fadeIn 1s forwards;
+}
+
+/* Responsive */
+@media (max-width: 500px) {
+    h1 { font-size: 1.5rem; }
+    .welcome-animation { font-size: 2rem; }
+    input[type="text"] { width: 200px; }
+}
+
