+++ public/css/productos.css
@@ -1,0 +1,341 @@
+/* ==== ESTILOS GENERALES ==== */
+body {
+  margin: 0;
+  font-family: "Poppins", sans-serif;
+  background-color: #121212;
+  color: #fff;
+}
+
+header {
+  position:relative;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 1rem 2rem;
+  background-color: #1f1f1f;
+  border-bottom: 2px solid #333;
+}
+
+.logo {
+  height: 100px;
+}
+
+header h1 {
+  font-size: 1.5rem;
+}
+
+nav a {
+  color: #fff;
+  text-decoration: none;
+  margin-left: 1rem;
+  font-weight: bold;
+  transition: color 0.3s;
+}
+
+nav a:hover {
+  color: #00bcd4;
+}
+
+/* ==== GRID DE PRODUCTOS ==== */
+.productos-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(230px, 1fr));
+  gap: 1.5rem;
+  padding: 1.5rem;
+}
+
+/* ==== TARJETA DE PRODUCTO ==== */
+.producto-card {
+  background-color: #1e1e1e;
+  border-radius: 12px;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.3);
+  overflow: hidden;
+  text-align: center;
+  transition: transform 0.2s, box-shadow 0.2s;
+  padding-bottom: 1rem;
+}
+
+.producto-card:hover {
+  transform: scale(1.03);
+  box-shadow: 0 6px 16px rgba(0, 0, 0, 0.4);
+}
+
+.producto-card img {
+  width: 100%;
+  height: 180px;
+  object-fit: cover;
+}
+
+.producto-card h3 {
+  margin: 0.5rem 0;
+  font-size: 1.1rem;
+}
+
+.categoria {
+  color: #bbb;
+  font-size: 0.9rem;
+}
+
+.precio {
+  font-size: 1.2rem;
+  font-weight: bold;
+  color: #00bcd4;
+}
+
+/* ==== BOTÓN ==== */
+.btn-agregar {
+  background-color: #00bcd4;
+  border: none;
+  color: white;
+  padding: 0.6rem 1rem;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: background 0.3s;
+  font-weight: bold;
+}
+
+.btn-agregar:hover {
+  background-color: #0097a7;
+}
+
+/* ==== FOOTER ==== */
+footer {
+  text-align: center;
+  padding: 1rem;
+  background-color: #1f1f1f;
+  color: #999;
+  border-top: 2px solid #333;
+}
+footer a{
+  padding: 0.5rem;
+  color: #bbb8;
+  text-decoration: none;
+  
+}
+
+footer a img{
+  border-radius: 35%;
+
+}
+
+footer a:hover{
+  text-decoration: underline;
+  color: #ddd;
+}
+
+
+
+/* Estilo para el logo */
+h1 {
+    width: auto;          /* tamaño aproximado de un h1 */
+    height: auto;         /* mantiene la proporción */
+    display: block;
+    margin: 20px auto;    /* lo centra verticalmente con margen */
+}
+
+/* Si querés alinearlo al centro en la parte superior */
+header .h1 {
+    display: block;
+    margin: 10px auto;
+}
+button {
+  background-color: transparent;
+  border: none; 
+  color: inherit; 
+  cursor: pointer; 
+}
+/* Ubica el menú arriba a la izquierda */
+.menu {
+  right: 1400px;
+  position: relative;
+  display: inline-block;
+  z-index: 1;
+}
+
+/* Estilo del botón */
+.menu-button {
+  color: white;
+  padding: 10px 20px;
+  cursor: pointer;
+  border: none;
+  border-radius: 5px;
+  font-size: 16px;
+}
+
+/* Contenedor del menú desplegable */
+.dropdown {
+  display: none;
+  position: absolute;
+  background-color: #f1f1f1;
+  min-width: 160px;
+  box-shadow: 0px 8px 16px rgba(0,0,0,0.2);
+}
+
+/* Estilo de los enlaces del menú */
+.dropdown a {
+  color: black;
+  padding: 12px 16px;
+  text-decoration: none;
+  display: block;
+}
+
+/* Hover sobre los enlaces */
+.dropdown a:hover {
+  background-color: #ddd;
+}
+
+/* Mostrar el menú al pasar el mouse */
+.menu:hover .dropdown {
+  display: block;
+}
+
+
+/* ============================= */
+/* MEDIA QUERIES PARA TABLETS    */
+/* ============================= */
+@media (max-width: 1024px) {
+
+  #logoBtn{
+    width: auto;
+    height: 0;
+  }
+
+  header {
+    flex-wrap: wrap;
+    gap: 1rem;
+  }
+
+  header h1 img {
+    width: 450px;
+    height: auto;
+  }
+
+
+  nav a {
+    margin-left: 0.5rem;
+    font-size: 0.9rem;
+  }
+
+  .menu {
+    position: relative;
+    right: 0;
+  }
+  .menu:hover .dropdown {
+    display: inline-flex;
+  }
+
+  .productos-grid {
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
+
+/* ============================= */
+/* MEDIA QUERIES PARA CELULARES GRANDES */
+/* ============================= */
+@media (max-width: 768px) {
+
+  #logoBtn{
+    width: auto;
+    height: 0;
+  }
+
+  header {
+    flex-direction: column;
+    text-align: center;
+  }
+
+
+  .menu {
+    order: 2;
+  }
+
+  .menu:hover .dropdown {
+  display: inline-flex;
+  }
+
+
+  nav {
+    order: 3;
+  }
+
+  .productos-grid {
+    grid-template-columns: repeat(2, 1fr);
+  }
+
+  .producto-card img {
+    height: 200px;
+    object-fit: cover;
+  }
+}
+
+/* ============================= */
+/* MEDIA QUERIES PARA CELULARES CHICOS */
+/* ============================= */
+@media (max-width: 480px) {
+
+  #logoBtn{
+    width: auto;
+    height: 0;
+  }
+
+
+  nav {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+  }
+
+  nav a {
+    margin-left: 0;
+  }
+
+  .dropdown {
+  display: none;
+  position: absolute;
+  background-color: #f1f1f1;
+  min-width: 100px;
+  box-shadow: 0px 8px 16px rgba(0,0,0,0.2);
+  }
+
+  .dropdown a {
+  color: black;
+  padding: 5px 5px;
+  text-decoration: none;
+  display: block;
+}
+
+  .productos-grid {
+    grid-template-columns: 1fr;
+  }
+
+  .producto-card img {
+    height: 180px;
+    object-fit: scale-down;
+  }
+
+  header h1 img {
+    width: 200px;
+    height: auto;
+  }
+}
+
+body.oscuro {
+  background-color: #121212;
+  color: #f5f5f5;
+}
+
+body.oscuro header {
+  background-color: #1e1e1e;
+}
+
+body.oscuro .btn-vaciar,
+body.oscuro button {
+  background-color: #333;
+  color: #f5f5f5;
+  border: 1px solid #555;
+}
+
+body.oscuro .modal-contenido {
+  background-color: #222;
+  color: #fff;
+}
