+++ public/index.html
@@ -1,0 +1,69 @@
+<!DOCTYPE html>
+<html lang="es">
+<head>
+    <meta charset="UTF-8">
+    <title>TP1 - Bienvenida</title>
+    <link rel="stylesheet" href="css/bienvenida.css">
+    <link rel="icon" href="favicon.ico">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+</head>
+<body>
+    <!-- Animación de bienvenida -->
+    <div class="welcome-animation" id="welcome">
+        Bienvenido a TP1
+    </div>
+
+    <!-- Contenedor principal que aparece después -->
+    <div class="main-content" id="mainContent">
+        <img src="logo.png" alt="Logo" class="logo">
+        <h1><img src="nombre.png" alt="Nombre Aplicación"></h1>
+
+        <!-- Formulario para ingresar nombre -->
+        <div class="name-section">
+            <input type="text" id="username" placeholder="Ingresa tu nombre">
+            <button id="continueBtn">Continuar</button>
+        </div>
+
+        <!-- Barra de navegación mínima -->
+        <nav class="nav-footer">
+            <button id="adminLoginBtn">Login Administrador</button>
+        </nav>
+    </div>
+
+    <script>
+        // Animación bienvenida
+        window.addEventListener('load', () => {
+            const welcome = document.getElementById('welcome');
+            const mainContent = document.getElementById('mainContent');
+            
+            setTimeout(() => {
+                welcome.style.opacity = 0;
+                welcome.style.transform = 'scale(1.2)';
+                setTimeout(() => {
+                    welcome.style.display = 'none';
+                    mainContent.classList.add('fade-in');
+                }, 800);
+            }, 2000);
+        });
+
+         // Redirección al hacer clic en "Continuar"
+        const continueBtn = document.getElementById('continueBtn');
+        continueBtn.addEventListener('click', () => {
+            const username = document.getElementById('username').value.trim();
+            if (username === "") {
+                alert("Por favor, ingresá tu nombre antes de continuar");
+            } else {
+                // Redirige a la página de productos
+                window.location.href = "productos.html";
+            }
+        });
+
+        // Botón de login administrador
+        const adminLoginBtn = document.getElementById('adminLoginBtn');
+        adminLoginBtn.addEventListener('click', () => {
+            window.location.href = "admin/login.html";
+        });
+    </script>
+
+</body>
+</html>
